Moved sim files into separate package.
